LAB_3/LAB_3/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
opencv-python-headless
numpy
python-dotenv